Renaming beacon/log.py to beacon/log_beacon.py to avoid conflicts when importing logging during tests.
