+++ salt/beacons/log_beacon.py
